Refactor methods
