+++ core/src/main/java/com/portto/sdk/core/method/RequestAccountMethod.kt
@@ -1,7 +1,9 @@
 package com.portto.sdk.core.method
 
+import android.net.Uri
 import com.portto.sdk.core.Blockchain
 import com.portto.sdk.wallet.BloctoSDKError
+import com.portto.sdk.wallet.Const
 
 class RequestAccountMethod(
     blockchain: Blockchain,
@@ -11,4 +13,13 @@ class RequestAccountMethod(
 
     override val name: String
         get() = "request_account"
+
+    override fun handleCallback(uri: Uri) {
+        val address = uri.getQueryParameter(Const.KEY_ADDRESS)
+        if (address.isNullOrEmpty()) {
+            onError(BloctoSDKError.INVALID_RESPONSE)
+            return
+        }
+        onSuccess(address)
+    }
 }
