+++ core/src/test/java/com/portto/sdk/core/BloctoSDKTest.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import android.content.Intent
 import android.net.Uri
 import com.portto.sdk.core.method.RequestAccountMethod
-import com.portto.sdk.core.method.SignAndSendTransactionMethod
 import com.portto.sdk.wallet.BloctoSDKError
 import com.portto.sdk.wallet.Const
 import io.mockk.*
@@ -49,8 +48,6 @@ class BloctoSDKTest {
     private val onError: (BloctoSDKError) -> Unit = mockk(relaxed = true)
     private val appId = "57f397df-263c-4e97-b61f-15b67b9ce285"
     private val solAddress = "zJ9A5VfCFdUsXAxouniMbDPMuj8MrWBosDwoQA3D78j"
-    private val txHash =
-        "2ZTRtryBTAPkbgALqtX9zi9TuuHCxzPBXDVQicDNRdQ7rSNH9Pb1zvAbKoissLSQ4vUeeNn2FhYegYmCDaQY6Jhw"
 
     @Test(expected = NullPointerException::class)
     fun `test without app id setup`() {
@@ -67,7 +64,7 @@ class BloctoSDKTest {
         requestAccount()
 
         val successCallbackUri = Uri.Builder()
-            .scheme("blocto")
+            .scheme(Const.BLOCTO_SCHEME)
             .appendQueryParameter(Const.KEY_REQUEST_ID, requestId)
             .appendQueryParameter(Const.KEY_ADDRESS, solAddress)
             .build()
@@ -85,42 +82,7 @@ class BloctoSDKTest {
 
         val error = BloctoSDKError.USER_REJECTED
         val errorCallbackUri = Uri.Builder()
-            .scheme("blocto")
-            .appendQueryParameter(Const.KEY_REQUEST_ID, requestId)
-            .appendQueryParameter(Const.KEY_ERROR, error.message)
-            .build()
-
-        BloctoSDK.handleCallback(errorCallbackUri)
-
-        val errorSlot = slot<BloctoSDKError>()
-        verify { onError(capture(errorSlot)) }
-        assertEquals(error.message, errorSlot.captured.message)
-    }
-
-    @Test
-    fun `test sign and send transaction with success callback`() {
-        signAndSendTransaction()
-
-        val successCallbackUri = Uri.Builder()
-            .scheme("blocto")
-            .appendQueryParameter(Const.KEY_REQUEST_ID, requestId)
-            .appendQueryParameter(Const.KEY_TX_HASH, txHash)
-            .build()
-
-        BloctoSDK.handleCallback(successCallbackUri)
-
-        val txHashSlot = slot<String>()
-        verify { onSuccess(capture(txHashSlot)) }
-        assertEquals(txHash, txHashSlot.captured)
-    }
-
-    @Test
-    fun `test sign and send transaction with error callback`() {
-        signAndSendTransaction()
-
-        val error = BloctoSDKError.USER_NOT_MATCH
-        val errorCallbackUri = Uri.Builder()
-            .scheme("blocto")
+            .scheme(Const.BLOCTO_SCHEME)
             .appendQueryParameter(Const.KEY_REQUEST_ID, requestId)
             .appendQueryParameter(Const.KEY_ERROR, error.message)
             .build()
@@ -155,70 +117,5 @@ class BloctoSDKTest {
         assertEquals(appId, capturedData?.getQueryParameter(Const.KEY_APP_ID))
         assertEquals(requestId, capturedData?.getQueryParameter(Const.KEY_REQUEST_ID))
         assertEquals(blockchain.value, capturedData?.getQueryParameter(Const.KEY_BLOCKCHAIN))
-    }
-
-    private fun signAndSendTransaction() {
-        val intentSlot = slot<Intent>()
-        val blockchain = Blockchain.SOLANA
-        val message = "test_message"
-        val isInvokeWrapped = true
-        val publicKeySignaturePairs = mapOf(
-            "public_key_1" to "signature_1",
-            "public_key_2" to "signature_2"
-        )
-        val appendTx = mapOf(
-            "id_1" to "tx_1",
-            "id_2" to "tx_2"
-        )
-
-        BloctoSDK.init(appId)
-
-        val method = spyk(
-            SignAndSendTransactionMethod(
-                fromAddress = solAddress,
-                message = message,
-                isInvokeWrapped = true,
-                publicKeySignaturePairs = publicKeySignaturePairs,
-                appendTx = appendTx,
-                blockchain = blockchain,
-                onSuccess = onSuccess,
-                onError = onError
-            )
-        )
-        BloctoSDK.send(context, method)
-
-        verify { context.startActivity(capture(intentSlot)) }
-
-        val capturedData = intentSlot.captured.data
-        assertEquals(Const.bloctoAuthority(BloctoSDK.debug), capturedData?.authority)
-        assertEquals(method.name, capturedData?.getQueryParameter(Const.KEY_METHOD))
-        assertEquals(appId, capturedData?.getQueryParameter(Const.KEY_APP_ID))
-        assertEquals(requestId, capturedData?.getQueryParameter(Const.KEY_REQUEST_ID))
-        assertEquals(blockchain.value, capturedData?.getQueryParameter(Const.KEY_BLOCKCHAIN))
-        assertEquals(message, capturedData?.getQueryParameter(Const.KEY_MESSAGE))
-        assertEquals(
-            isInvokeWrapped,
-            capturedData?.getQueryParameter(Const.KEY_IS_INVOKE_WRAPPED)?.toBoolean()
-        )
-        assertEquals(
-            publicKeySignaturePairs,
-            capturedData
-                ?.queryParameterNames
-                ?.filter { it.startsWith(Const.KEY_PUBLIC_KEY_SIGNATURE_PAIRS) }
-                ?.associate {
-                    val publicKey = it.substringAfter("[").substringBefore("]")
-                    publicKey to (capturedData.getQueryParameter(it) ?: "")
-                }
-        )
-        assertEquals(
-            appendTx,
-            capturedData
-                ?.queryParameterNames
-                ?.filter { it.startsWith(Const.KEY_APPEND_TX) }
-                ?.associate {
-                    val hash = it.substringAfter("[").substringBefore("]")
-                    hash to (capturedData.getQueryParameter(it) ?: "")
-                }
-        )
     }
 }
