+++ solana/src/main/java/com/portto/sdk/solana/Solana.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.annotation.WorkerThread
 import com.portto.sdk.core.*
 import com.portto.sdk.core.method.RequestAccountMethod
-import com.portto.sdk.core.method.SignAndSendTransactionMethod
+import com.portto.sdk.solana.method.SignAndSendTransactionMethod
 import com.portto.sdk.wallet.BloctoSDKError
 import com.portto.solana.web3.AccountMeta
 import com.portto.solana.web3.Message
