+++ solana/src/main/java/com/portto/sdk/solana/method/SignAndSendTransactionMethod.kt
@@ -1,7 +1,8 @@
-package com.portto.sdk.core.method
+package com.portto.sdk.solana.method
 
 import android.net.Uri
 import com.portto.sdk.core.Blockchain
+import com.portto.sdk.core.method.Method
 import com.portto.sdk.wallet.BloctoSDKError
 import com.portto.sdk.wallet.Const
 
@@ -18,6 +19,15 @@ class SignAndSendTransactionMethod(
 
     override val name: String
         get() = "sign_and_send_transaction"
+
+    override fun handleCallback(uri: Uri) {
+        val txHash = uri.getQueryParameter(Const.KEY_TX_HASH)
+        if (txHash.isNullOrEmpty()) {
+            onError(BloctoSDKError.INVALID_RESPONSE)
+            return
+        }
+        onSuccess(txHash)
+    }
 
     override fun encodeToUri(authority: String, appId: String, requestId: String): Uri.Builder {
         return super.encodeToUri(authority, appId, requestId)
